Fix layout rendering in Storybook
